+++ src/api/models/__init__.py
@@ -9,3 +9,4 @@ from src.api.models.garages.locations import Locations
 from src.api.models.licence_plates import LicencePlates
 from src.api.models.parking_lots import ParkingLots
 from src.api.models.settings import Settings
+from src.api.models.reservations import Reservations
