+++ src/api/seeds/fixtures/0002_locations_fixture.json
@@ -1,0 +1,142 @@
+[
+    {
+        "model": "api.locations",
+        "pk": 1,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "OVL",
+            "municipality": "Membruggen",
+            "post_code": 3376,
+            "street": "Rebeccastraat",
+            "number": 114,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    },
+    {
+        "model": "api.locations",
+        "pk": 2,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "ANT",
+            "municipality": "Soiron",
+            "post_code": 2065,
+            "street": "Jessepad",
+            "number": 147,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    },
+    {
+        "model": "api.locations",
+        "pk": 3,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "LIM",
+            "municipality": "Joncret",
+            "post_code": 2486,
+            "street": "Amysingel",
+            "number": 109,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    },
+    {
+        "model": "api.locations",
+        "pk": 4,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "LIM",
+            "municipality": "Seilles",
+            "post_code": 3299,
+            "street": "Staceystraat",
+            "number": 67,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    },
+    {
+        "model": "api.locations",
+        "pk": 5,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "ANT",
+            "municipality": "Moorsele",
+            "post_code": 3804,
+            "street": "Tinasteeg",
+            "number": 160,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    },
+    {
+        "model": "api.locations",
+        "pk": 6,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "OVL",
+            "municipality": "Keumi\u00e9e",
+            "post_code": 3438,
+            "street": "Brianbaan",
+            "number": 154,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    },
+    {
+        "model": "api.locations",
+        "pk": 7,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "ANT",
+            "municipality": "Sint-Lambrechts-Herk",
+            "post_code": 2001,
+            "street": "Herbertring",
+            "number": 91,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    },
+    {
+        "model": "api.locations",
+        "pk": 8,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "OVL",
+            "municipality": "Ouwegem",
+            "post_code": 3630,
+            "street": "Edwardstraat",
+            "number": 7,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    },
+    {
+        "model": "api.locations",
+        "pk": 9,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "OVL",
+            "municipality": "Tienen",
+            "post_code": 2703,
+            "street": "Katherinelei",
+            "number": 55,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    },
+    {
+        "model": "api.locations",
+        "pk": 10,
+        "fields": {
+            "country": "Belgi\u00eb",
+            "province": "LIM",
+            "municipality": "Sainte-Ode",
+            "post_code": 2264,
+            "street": "Crystallei",
+            "number": 173,
+            "updated_at": "2022-11-16T05:11:57.520272+00:00",
+            "created_at": "2022-11-16T05:11:57.520290+00:00"
+        }
+    }
+]
