+++ src/api/seeds/fixtures/0003_garage_settings_fixture.json
@@ -1,0 +1,122 @@
+[
+    {
+        "model": "api.garageSettings",
+        "pk": 1,
+        "fields": {
+            "max_height": 2.3,
+            "location": 1,
+            "max_width": 2.4,
+            "max_handicapped_lots": 6,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    },
+    {
+        "model": "api.garageSettings",
+        "pk": 2,
+        "fields": {
+            "max_height": 2.1,
+            "location": 2,
+            "max_width": 2.1,
+            "max_handicapped_lots": 8,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    },
+    {
+        "model": "api.garageSettings",
+        "pk": 3,
+        "fields": {
+            "max_height": 2.2,
+            "location": 3,
+            "max_width": 2.2,
+            "max_handicapped_lots": 4,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    },
+    {
+        "model": "api.garageSettings",
+        "pk": 4,
+        "fields": {
+            "max_height": 2.1,
+            "location": 4,
+            "max_width": 2.3,
+            "max_handicapped_lots": 2,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    },
+    {
+        "model": "api.garageSettings",
+        "pk": 5,
+        "fields": {
+            "max_height": 2.0,
+            "location": 5,
+            "max_width": 2.1,
+            "max_handicapped_lots": 6,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    },
+    {
+        "model": "api.garageSettings",
+        "pk": 6,
+        "fields": {
+            "max_height": 2.2,
+            "location": 6,
+            "max_width": 1.9,
+            "max_handicapped_lots": 4,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    },
+    {
+        "model": "api.garageSettings",
+        "pk": 7,
+        "fields": {
+            "max_height": 2.2,
+            "location": 7,
+            "max_width": 2.2,
+            "max_handicapped_lots": 5,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    },
+    {
+        "model": "api.garageSettings",
+        "pk": 8,
+        "fields": {
+            "max_height": 2.1,
+            "location": 8,
+            "max_width": 2.1,
+            "max_handicapped_lots": 5,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    },
+    {
+        "model": "api.garageSettings",
+        "pk": 9,
+        "fields": {
+            "max_height": 1.8,
+            "location": 9,
+            "max_width": 2.2,
+            "max_handicapped_lots": 7,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    },
+    {
+        "model": "api.garageSettings",
+        "pk": 10,
+        "fields": {
+            "max_height": 1.9,
+            "location": 10,
+            "max_width": 1.9,
+            "max_handicapped_lots": 4,
+            "created_at": "2022-11-16T05:11:57.520642+00:00",
+            "updated_at": "2022-11-16T05:11:57.520632+00:00"
+        }
+    }
+]
