+++ src/api/seeds/fixtures/0006_licence_plates_fixture.json
@@ -6,8 +6,8 @@
             "user": 1,
             "licence_plate": "1ABC123",
             "garage": 1,
-            "created_at": "2022-11-15T20:50:19.864396+01:00",
-            "updated_at": "2022-11-15T20:50:19.864348+01:00"
+            "created_at": "2022-11-16T05:11:57.522643+00:00",
+            "updated_at": "2022-11-16T05:11:57.522631+00:00"
         }
     },
     {
@@ -17,8 +17,8 @@
             "user": 1,
             "licence_plate": "2JIO489",
             "garage": null,
-            "created_at": "2022-11-15T20:50:19.864396+01:00",
-            "updated_at": "2022-11-15T20:50:19.864348+01:00"
+            "created_at": "2022-11-16T05:11:57.522643+00:00",
+            "updated_at": "2022-11-16T05:11:57.522631+00:00"
         }
     },
     {
@@ -28,8 +28,8 @@
             "user": 2,
             "licence_plate": "1JFK908",
             "garage": 3,
-            "created_at": "2022-11-15T20:50:19.864396+01:00",
-            "updated_at": "2022-11-15T20:50:19.864348+01:00"
+            "created_at": "2022-11-16T05:11:57.522643+00:00",
+            "updated_at": "2022-11-16T05:11:57.522631+00:00"
         }
     },
     {
@@ -39,8 +39,8 @@
             "user": 2,
             "licence_plate": "8FJO108",
             "garage": 4,
-            "created_at": "2022-11-15T20:50:19.864396+01:00",
-            "updated_at": "2022-11-15T20:50:19.864348+01:00"
+            "created_at": "2022-11-16T05:11:57.522643+00:00",
+            "updated_at": "2022-11-16T05:11:57.522631+00:00"
         }
     },
     {
@@ -50,8 +50,8 @@
             "user": 3,
             "licence_plate": "2AZP367",
             "garage": 5,
-            "created_at": "2022-11-15T20:50:19.864396+01:00",
-            "updated_at": "2022-11-15T20:50:19.864348+01:00"
+            "created_at": "2022-11-16T05:11:57.522643+00:00",
+            "updated_at": "2022-11-16T05:11:57.522631+00:00"
         }
     },
     {
@@ -61,8 +61,8 @@
             "user": 4,
             "licence_plate": "1KPX789",
             "garage": null,
-            "created_at": "2022-11-15T20:50:19.864396+01:00",
-            "updated_at": "2022-11-15T20:50:19.864348+01:00"
+            "created_at": "2022-11-16T05:11:57.522643+00:00",
+            "updated_at": "2022-11-16T05:11:57.522631+00:00"
         }
     },
     {
@@ -72,8 +72,8 @@
             "user": 5,
             "licence_plate": "1FKO392",
             "garage": null,
-            "created_at": "2022-11-15T20:50:19.864396+01:00",
-            "updated_at": "2022-11-15T20:50:19.864348+01:00"
+            "created_at": "2022-11-16T05:11:57.522643+00:00",
+            "updated_at": "2022-11-16T05:11:57.522631+00:00"
         }
     }
 ]
