+++ src/api/seeds/fixtures/0007_prices_fixture.json
@@ -1,0 +1,242 @@
+[
+    {
+        "model": "api.prices",
+        "pk": 1,
+        "fields": {
+            "garage": 1,
+            "price_string": "3 uur parkeren",
+            "price": 6.4,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 2,
+        "fields": {
+            "garage": 1,
+            "price_string": "10 uur parkeren",
+            "price": 9.9,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 3,
+        "fields": {
+            "garage": 2,
+            "price_string": "9 uur parkeren",
+            "price": 6.8,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 4,
+        "fields": {
+            "garage": 2,
+            "price_string": "4 uur parkeren",
+            "price": 10.3,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 5,
+        "fields": {
+            "garage": 3,
+            "price_string": "4 uur parkeren",
+            "price": 4.3,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 6,
+        "fields": {
+            "garage": 3,
+            "price_string": "7 uur parkeren",
+            "price": 10.6,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 7,
+        "fields": {
+            "garage": 4,
+            "price_string": "9 uur parkeren",
+            "price": 8.2,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 8,
+        "fields": {
+            "garage": 4,
+            "price_string": "8 uur parkeren",
+            "price": 10.9,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 9,
+        "fields": {
+            "garage": 5,
+            "price_string": "3 uur parkeren",
+            "price": 14.2,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 10,
+        "fields": {
+            "garage": 5,
+            "price_string": "9 uur parkeren",
+            "price": 10.3,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 11,
+        "fields": {
+            "garage": 6,
+            "price_string": "5 uur parkeren",
+            "price": 11.5,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 12,
+        "fields": {
+            "garage": 6,
+            "price_string": "2 uur parkeren",
+            "price": 14.3,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 13,
+        "fields": {
+            "garage": 7,
+            "price_string": "5 uur parkeren",
+            "price": 5.2,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 14,
+        "fields": {
+            "garage": 7,
+            "price_string": "9 uur parkeren",
+            "price": 13.7,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 15,
+        "fields": {
+            "garage": 8,
+            "price_string": "9 uur parkeren",
+            "price": 11.8,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 16,
+        "fields": {
+            "garage": 8,
+            "price_string": "9 uur parkeren",
+            "price": 4.1,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 17,
+        "fields": {
+            "garage": 9,
+            "price_string": "8 uur parkeren",
+            "price": 5.2,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 18,
+        "fields": {
+            "garage": 9,
+            "price_string": "4 uur parkeren",
+            "price": 6.8,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 19,
+        "fields": {
+            "garage": 10,
+            "price_string": "7 uur parkeren",
+            "price": 9.6,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    },
+    {
+        "model": "api.prices",
+        "pk": 20,
+        "fields": {
+            "garage": 10,
+            "price_string": "1 uur parkeren",
+            "price": 14.1,
+            "valuta": "EUR",
+            "created_at": "2022-11-16T05:11:57.522863+00:00",
+            "updated_at": "2022-11-16T05:11:57.522854+00:00"
+        }
+    }
+]
