+++ src/api/seeds/fixtures/0008_opening_hours_fixture.json
@@ -1,0 +1,262 @@
+[
+    {
+        "model": "api.openingHours",
+        "pk": 1,
+        "fields": {
+            "garage": 1,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 2,
+        "fields": {
+            "garage": 1,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 3,
+        "fields": {
+            "garage": 2,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 4,
+        "fields": {
+            "garage": 2,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 5,
+        "fields": {
+            "garage": 3,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 6,
+        "fields": {
+            "garage": 3,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 7,
+        "fields": {
+            "garage": 4,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 8,
+        "fields": {
+            "garage": 4,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 9,
+        "fields": {
+            "garage": 5,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 10,
+        "fields": {
+            "garage": 5,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 11,
+        "fields": {
+            "garage": 6,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 12,
+        "fields": {
+            "garage": 6,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 13,
+        "fields": {
+            "garage": 7,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 14,
+        "fields": {
+            "garage": 7,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 15,
+        "fields": {
+            "garage": 8,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 16,
+        "fields": {
+            "garage": 8,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 17,
+        "fields": {
+            "garage": 9,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 18,
+        "fields": {
+            "garage": 9,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 19,
+        "fields": {
+            "garage": 10,
+            "from_day": 0,
+            "to_day": 5,
+            "from_hour": "00:00:00",
+            "to_hour": "23:59:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    },
+    {
+        "model": "api.openingHours",
+        "pk": 20,
+        "fields": {
+            "garage": 10,
+            "from_day": 6,
+            "to_day": 6,
+            "from_hour": "05:00:00",
+            "to_hour": "22:00:00",
+            "updated_at": "2022-11-16T05:11:57.523207+00:00",
+            "created_at": "2022-11-16T05:11:57.523216+00:00"
+        }
+    }
+]
