+++ src/api/seeds/fixtures/0009_reservations_fixture.json
@@ -1,0 +1,132 @@
+[
+    {
+        "model": "api.reservations",
+        "pk": 1,
+        "fields": {
+            "garage": 8,
+            "user": 2,
+            "parking_lot": 59,
+            "from_date": "2022-12-21T02:00:00+00:00",
+            "to_date": "2022-12-22T19:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    },
+    {
+        "model": "api.reservations",
+        "pk": 2,
+        "fields": {
+            "garage": 5,
+            "user": 3,
+            "parking_lot": 69,
+            "from_date": "2022-12-20T02:00:00+00:00",
+            "to_date": "2022-12-21T18:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    },
+    {
+        "model": "api.reservations",
+        "pk": 3,
+        "fields": {
+            "garage": 9,
+            "user": 4,
+            "parking_lot": 26,
+            "from_date": "2022-12-20T04:00:00+00:00",
+            "to_date": "2022-12-21T19:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    },
+    {
+        "model": "api.reservations",
+        "pk": 4,
+        "fields": {
+            "garage": 9,
+            "user": 2,
+            "parking_lot": 8,
+            "from_date": "2022-12-21T00:00:00+00:00",
+            "to_date": "2022-12-22T16:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    },
+    {
+        "model": "api.reservations",
+        "pk": 5,
+        "fields": {
+            "garage": 9,
+            "user": 3,
+            "parking_lot": 94,
+            "from_date": "2022-12-20T05:00:00+00:00",
+            "to_date": "2022-12-22T21:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    },
+    {
+        "model": "api.reservations",
+        "pk": 6,
+        "fields": {
+            "garage": 5,
+            "user": 4,
+            "parking_lot": 22,
+            "from_date": "2022-12-21T06:00:00+00:00",
+            "to_date": "2022-12-21T17:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    },
+    {
+        "model": "api.reservations",
+        "pk": 7,
+        "fields": {
+            "garage": 4,
+            "user": 2,
+            "parking_lot": 52,
+            "from_date": "2022-12-20T01:00:00+00:00",
+            "to_date": "2022-12-22T16:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    },
+    {
+        "model": "api.reservations",
+        "pk": 8,
+        "fields": {
+            "garage": 3,
+            "user": 1,
+            "parking_lot": 77,
+            "from_date": "2022-12-21T02:00:00+00:00",
+            "to_date": "2022-12-22T15:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    },
+    {
+        "model": "api.reservations",
+        "pk": 9,
+        "fields": {
+            "garage": 3,
+            "user": 1,
+            "parking_lot": 22,
+            "from_date": "2022-12-20T04:00:00+00:00",
+            "to_date": "2022-12-21T17:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    },
+    {
+        "model": "api.reservations",
+        "pk": 10,
+        "fields": {
+            "garage": 8,
+            "user": 4,
+            "parking_lot": 83,
+            "from_date": "2022-12-21T03:00:00+00:00",
+            "to_date": "2022-12-21T21:00:00+00:00",
+            "updated_at": "2022-11-16T05:11:57.523691+00:00",
+            "created_at": "2022-11-16T05:11:57.523696+00:00"
+        }
+    }
+]
