--- src/api/seeds/fixtures/001_users_fixture.json
@@ -1,72 +1,0 @@
-[
-  {
-    "model": "users.user",
-    "pk": 1,
-    "fields": {
-      "email": "john.doe@po.be",
-      "first_name": "John",
-      "last_name": "Doe",
-      "password": "pbkdf2_sha256$390000$8oBh06haR0eVAqYsOCXxYG$g0Q67b08ZZAPxGJQlW+vA9EaLlaK4o+KwHeQ9wVRPwk=",
-      "is_active": true,
-      "role": 2,
-      "created_at": "2022-11-07T19:03:57.345475+00:00",
-      "updated_at": "2022-11-07T19:03:57.345475+00:00"
-    }
-  },
-  {
-    "model": "users.user",
-    "pk": 2,
-    "fields": {
-      "email": "albert.einstein@po.be",
-      "first_name": "Albert",
-      "last_name": "Einstein",
-      "password": "pbkdf2_sha256$390000$YdB4dafSO5YtsRBDi9GSPb$PM5gcvNK9mQpNh3ugfZMcQ2hTShasSfwHjvEGmSXvGo=",
-      "is_active": true,
-      "role": 3,
-      "created_at": "2022-11-06T19:03:57.345475+00:00",
-      "updated_at": "2022-11-06T19:03:57.345475+00:00"
-    }
-  },
-  {
-    "model": "users.user",
-    "pk": 3,
-    "fields": {
-      "email": "developer@po.be",
-      "first_name": "Grace",
-      "last_name": "Hopper",
-      "password": "pbkdf2_sha256$390000$MiCZ1QI7EZw0OIzZwR8rRV$SFJZDhWiL8M5UrKUbeL2+GI1UYhBzWgWiI6nsLKZPLk=",
-      "is_active": true,
-      "role": 3,
-      "created_at": "2022-11-05T19:03:57.345475+00:00",
-      "updated_at": "2022-11-05T19:03:57.345475+00:00"
-    }
-  },
-  {
-    "model": "users.user",
-    "pk": 4,
-    "fields": {
-      "email": "ernest.Hemmingway@po.be",
-      "first_name": "Ernest",
-      "last_name": "Hemmingway",
-      "password": "pbkdf2_sha256$390000$x8NC2Ubqx0cmTMTZJzomId$AZCkVmPhZbQwIv2zMcAoDGFtL0UkKnX8pm0b9qsd+I8=",
-      "is_active": true,
-      "role": 2,
-      "created_at": "2022-11-07T19:03:57.345475+00:00",
-      "updated_at": "2022-11-07T19:03:57.345475+00:00"
-    }
-  },
-  {
-    "model": "users.user",
-    "pk": 5,
-    "fields": {
-      "email": "isaak.newton@po.be",
-      "first_name": "Isaak",
-      "last_name": "Newton",
-      "password": "pbkdf2_sha256$390000$CxijX6XDh9kbKlp5JmKtPi$hS/Wi6zUUl5IxGFJ9Tze0kdGHDZ4I0PpDkcgU+DjhKg=",
-      "is_active": true,
-      "role": 1,
-      "created_at": "2022-11-07T19:03:57.345475+00:00",
-      "updated_at": "2022-11-07T19:03:57.345475+00:00"
-    }
-  }
-]
