--- src/api/seeds/fixtures/002_garages_fixture.json
@@ -1,102 +1,0 @@
-[
-    {
-        "model": "api.garages",
-        "pk": 1,
-        "fields": {
-            "owner": 1,
-            "name": "QPark Hasselt",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    },
-    {
-        "model": "api.garages",
-        "pk": 2,
-        "fields": {
-            "owner": 1,
-            "name": "QPark Leuven",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    },
-    {
-        "model": "api.garages",
-        "pk": 3,
-        "fields": {
-            "owner": 2,
-            "name": "QPark Aalter",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    },
-    {
-        "model": "api.garages",
-        "pk": 4,
-        "fields": {
-            "owner": 2,
-            "name": "QPark Brugge",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    },
-    {
-        "model": "api.garages",
-        "pk": 5,
-        "fields": {
-            "owner": 3,
-            "name": "QPark Herstappe",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    },
-    {
-        "model": "api.garages",
-        "pk": 6,
-        "fields": {
-            "owner": 3,
-            "name": "QPark Oostduinkerke",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    },
-    {
-        "model": "api.garages",
-        "pk": 7,
-        "fields": {
-            "owner": 4,
-            "name": "QPark Antwerpen",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    },
-    {
-        "model": "api.garages",
-        "pk": 8,
-        "fields": {
-            "owner": 4,
-            "name": "QPark Oostende",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    },
-    {
-        "model": "api.garages",
-        "pk": 9,
-        "fields": {
-            "owner": 5,
-            "name": "QPark Mariakerke",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    },
-    {
-        "model": "api.garages",
-        "pk": 10,
-        "fields": {
-            "owner": 5,
-            "name": "QPark Geel",
-            "updated_at": "2022-11-15T20:50:19.861657+01:00",
-            "created_at": "2022-11-15T20:50:19.861706+01:00"
-        }
-    }
-]
