--- src/api/seeds/fixtures/005_locations_fixture.json
@@ -1,32 +1,0 @@
-[
-    {
-        "model": "api.locations",
-        "pk": 1,
-        "fields": {
-            "garage_id": 1,
-            "country": "Belgi\u00eb",
-            "province": "LIM",
-            "municipality": "Hasselt",
-            "post_code": 3500,
-            "street": "Toekomststraat",
-            "number": 43,
-            "updated_at": "2022-11-15T20:50:19.864654+01:00",
-            "created_at": "2022-11-15T20:50:19.864667+01:00"
-        }
-    },
-    {
-        "model": "api.locations",
-        "pk": 2,
-        "fields": {
-            "garage_id": 2,
-            "country": "Belgi\u00eb",
-            "province": "LIM",
-            "municipality": "Lummen",
-            "post_code": 3000,
-            "street": "Sint-Annastraat",
-            "number": 590,
-            "updated_at": "2022-11-15T20:50:19.864654+01:00",
-            "created_at": "2022-11-15T20:50:19.864667+01:00"
-        }
-    }
-]
