--- src/api/seeds/fixtures/006_garage_settings_fixture.json
@@ -1,28 +1,0 @@
-[
-    {
-        "model": "api.garageSettings",
-        "pk": 1,
-        "fields": {
-            "garage_id": 1,
-            "max_height": 1.9,
-            "location": 1,
-            "max_width": 1.8,
-            "max_handicapped_lots": 10,
-            "created_at": "2022-11-15T20:50:19.864817+01:00",
-            "updated_at": "2022-11-15T20:50:19.864805+01:00"
-        }
-    },
-    {
-        "model": "api.garageSettings",
-        "pk": 2,
-        "fields": {
-            "garage_id": 2,
-            "max_height": 2.1,
-            "location": 2,
-            "max_width": 1.95,
-            "max_handicapped_lots": 12,
-            "created_at": "2022-11-15T20:50:19.864817+01:00",
-            "updated_at": "2022-11-15T20:50:19.864805+01:00"
-        }
-    }
-]
