+++ src/api/serializers/__init__.py
@@ -1,4 +1,5 @@
 from src.api.serializers.garages.garages_serializer import GaragesSerializer
+from src.api.serializers.garages.locations_serializer import LocationsSerializer
 from src.api.serializers.garages.garage_settings_serializer import (
     GarageSettingsSerializer,
 )
