+++ src/api/serializers/garages/garages_serializer.py
@@ -3,17 +3,18 @@ from rest_framework import serializers
 from src.api.models import Garages
 
 
-class LocationsSerializer(serializers.ModelSerializer):
-    postCode = serializers.BooleanField(source="post_code", read_only=True)
+class GaragesSerializer(serializers.ModelSerializer):
+    isFull = serializers.BooleanField(source="is_full", read_only=True)
+    unoccupiedLots = serializers.IntegerField(source="unoccupied_lots", read_only=True)
+    parkingLots = serializers.IntegerField(source="parking_lots", read_only=True)
 
     class Meta:
         model = Garages
         fields = [
             "id",
-            "country",
-            "province",
-            "municipality",
-            "postCode",
-            "street",
-            "number",
+            "owner",
+            "name",
+            "isFull",
+            "unoccupiedLots",
+            "parkingLots",
         ]
